fix type system, add functions, wildcard tests

+++ xpath3/parser_test.go
@@ -237,7 +237,7 @@ func TestParseAST(t *testing.T) {
 		require.True(t, ok, "expected RangeExpr, got %T", expr)
 		start, ok := re.Start.(xpath3.LiteralExpr)
 		require.True(t, ok)
-		require.Equal(t, float64(1), start.Value)
+		require.Equal(t, int64(1), start.Value)
 	})
 
 	t.Run("if expr structure", func(t *testing.T) {
